% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{add_tags}
\alias{add_tags}
\title{Attaches tags to an existing Elasticsearch domain}
\usage{
add_tags(ARN, TagList)
}
\arguments{
\item{ARN}{Specify the \code{ARN} for which you want to add the tags.}

\item{TagList}{List of \code{Tag} that need to be added for the Elasticsearch domain.}
}
\description{
Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-managedomains.html#es-managedomains-awsresorcetagging}{Tagging Amazon Elasticsearch Service Domains for more information.}
}
\examples{

}
