% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{delete_elasticsearch_domain}
\alias{delete_elasticsearch_domain}
\title{Permanently deletes the specified Elasticsearch domain and all of its data}
\usage{
delete_elasticsearch_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the Elasticsearch domain that you want to permanently delete.}
}
\description{
Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.
}
\section{Accepted Parameters}{
\preformatted{delete_elasticsearch_domain(
  DomainName = "string"
)
}
}

