% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{describe_elasticsearch_domain}
\alias{describe_elasticsearch_domain}
\title{Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN}
\usage{
describe_elasticsearch_domain(
  DomainName = "string"
)
}
\arguments{
\item{DomainName}{[required] The name of the Elasticsearch domain for which you want information.}
}
\description{
Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.
}
\examples{

}
