% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{describe_elasticsearch_domain_config}
\alias{describe_elasticsearch_domain_config}
\title{Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options}
\usage{
describe_elasticsearch_domain_config(DomainName)
}
\arguments{
\item{DomainName}{[required] The Elasticsearch domain that you want to get information about.}
}
\description{
Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.
}
\section{Accepted Parameters}{
\preformatted{describe_elasticsearch_domain_config(
  DomainName = "string"
)
}
}

