% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{describe_elasticsearch_domains}
\alias{describe_elasticsearch_domains}
\title{Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN}
\usage{
describe_elasticsearch_domains(
  DomainNames = list(
    "string"
  )
)
}
\arguments{
\item{DomainNames}{[required] The Elasticsearch domains for which you want information.}
}
\description{
Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.
}
\examples{

}
