% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{describe_elasticsearch_instance_type_limits}
\alias{describe_elasticsearch_instance_type_limits}
\title{Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion}
\usage{
describe_elasticsearch_instance_type_limits(InstanceType, ElasticsearchVersion,
  DomainName = NULL)
}
\arguments{
\item{InstanceType}{The instance type for an Elasticsearch cluster for which Elasticsearch \code{Limits Limits} are needed.}

\item{ElasticsearchVersion}{Version of Elasticsearch for which \code{Limits Limits} are needed.}

\item{DomainName}{DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch \code{Limits Limits} for existing domain.}
}
\description{
Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the \code{DomainName DomainName} to know what Limits are supported for modifying.
}
\examples{

}
