% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{get_compatible_elasticsearch_versions}
\alias{get_compatible_elasticsearch_versions}
\title{Returns a list of upgrade compatible Elastisearch versions}
\usage{
get_compatible_elasticsearch_versions(DomainName = NULL)
}
\arguments{
\item{DomainName}{}
}
\description{
Returns a list of upgrade compatible Elastisearch versions. You can optionally pass a \code{DomainName DomainName} to get all upgrade compatible Elasticsearch versions for that specific domain.
}
\section{Accepted Parameters}{
\preformatted{get_compatible_elasticsearch_versions(
  DomainName = "string"
)
}
}

