% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{get_upgrade_history}
\alias{get_upgrade_history}
\title{Retrieves the complete history of the last 10 upgrades that were performed on the domain}
\usage{
get_upgrade_history(DomainName, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required]}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Retrieves the complete history of the last 10 upgrades that were performed on the domain.
}
\section{Accepted Parameters}{
\preformatted{get_upgrade_history(
  DomainName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

