% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{get_upgrade_status}
\alias{get_upgrade_status}
\title{Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain}
\usage{
get_upgrade_status(
  DomainName = "string"
)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.
}
\examples{

}
