% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{list_domain_names}
\alias{list_domain_names}
\title{Returns the name of all Elasticsearch domains owned by the current user's account}
\usage{
list_domain_names()
}
\description{
Returns the name of all Elasticsearch domains owned by the current user's account.
}
