% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{list_elasticsearch_versions}
\alias{list_elasticsearch_versions}
\title{List all supported Elasticsearch versions}
\usage{
list_elasticsearch_versions(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.}

\item{NextToken}{}
}
\description{
List all supported Elasticsearch versions
}
\section{Accepted Parameters}{
\preformatted{list_elasticsearch_versions(
  MaxResults = 123,
  NextToken = "string"
)
}
}

