% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{list_tags}
\alias{list_tags}
\title{Returns all tags for the given Elasticsearch domain}
\usage{
list_tags(ARN)
}
\arguments{
\item{ARN}{[required] Specify the \code{ARN} for the Elasticsearch domain to which the tags are attached that you want to view.}
}
\description{
Returns all tags for the given Elasticsearch domain.
}
\section{Accepted Parameters}{
\preformatted{list_tags(
  ARN = "string"
)
}
}

