% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{purchase_reserved_elasticsearch_instance_offering}
\alias{purchase_reserved_elasticsearch_instance_offering}
\title{Allows you to purchase reserved Elasticsearch instances}
\usage{

  purchase_reserved_elasticsearch_instance_offering(ReservedElasticsearchInstanceOfferingId,
  ReservationName, InstanceCount = NULL)
}
\arguments{
\item{ReservedElasticsearchInstanceOfferingId}{The ID of the reserved Elasticsearch instance offering to purchase.}

\item{ReservationName}{A customer-specified identifier to track this reservation.}

\item{InstanceCount}{The number of Elasticsearch instances to reserve.}
}
\description{
Allows you to purchase reserved Elasticsearch instances.
}
\examples{

}
