% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{remove_tags}
\alias{remove_tags}
\title{Removes the specified set of tags from the specified Elasticsearch domain}
\usage{
remove_tags(ARN, TagKeys)
}
\arguments{
\item{ARN}{[required] Specifies the \code{ARN} for the Elasticsearch domain from which you want to delete the specified tags.}

\item{TagKeys}{[required] Specifies the \code{TagKey} list which you want to remove from the Elasticsearch domain.}
}
\description{
Removes the specified set of tags from the specified Elasticsearch domain.
}
\section{Accepted Parameters}{
\preformatted{remove_tags(
  ARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

