% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{start_elasticsearch_service_software_update}
\alias{start_elasticsearch_service_software_update}
\title{Schedules a service software update for an Amazon ES domain}
\usage{
start_elasticsearch_service_software_update(DomainName)
}
\arguments{
\item{DomainName}{The name of the domain that you want to update to the latest service software.}
}
\description{
Schedules a service software update for an Amazon ES domain.
}
\examples{

}
