% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticsearchservice_operations.R
\name{update_elasticsearch_domain_config}
\alias{update_elasticsearch_domain_config}
\title{Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances}
\usage{
update_elasticsearch_domain_config(DomainName,
  ElasticsearchClusterConfig = NULL, EBSOptions = NULL,
  SnapshotOptions = NULL, VPCOptions = NULL, CognitoOptions = NULL,
  AdvancedOptions = NULL, AccessPolicies = NULL,
  LogPublishingOptions = NULL)
}
\arguments{
\item{DomainName}{The name of the Elasticsearch domain that you are updating.}

\item{ElasticsearchClusterConfig}{The type and number of instances to instantiate for the domain cluster.}

\item{EBSOptions}{Specify the type and size of the EBS volume that you want to use.}

\item{SnapshotOptions}{Option to set the time, in UTC format, for the daily automated snapshot. Default value is \code{0} hours.}

\item{VPCOptions}{Options to specify the subnets and security groups for VPC endpoint. For more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-vpc.html#es-creating-vpc}{Creating a VPC} in \emph{VPC Endpoints for Amazon Elasticsearch Service Domains}}

\item{CognitoOptions}{Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-cognito-auth.html}{Amazon Cognito Authentication for Kibana}.}

\item{AdvancedOptions}{Modifies the advanced option to allow references to indices in an HTTP request body. Must be \code{false} when configuring access to individual sub-resources. By default, the value is \code{true}. See \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-createupdatedomains.html#es-createdomain-configure-advanced-options}{Configuration Advanced Options} for more information.}

\item{AccessPolicies}{IAM access policy as a JSON-formatted string.}

\item{LogPublishingOptions}{Map of \code{LogType} and \code{LogPublishingOption}, each containing options to publish a given type of Elasticsearch log.}
}
\description{
Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.
}
\examples{

}
